backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
motor>=3.3.2,<4.0.0
beanie>=1.26.0,<2.0.0
python-telegram-bot>=20.0
slowapi>=0.1.9
